OPTIMIZATION 16 bit INDICES

+++ Mesh.h
@@ -11,7 +11,7 @@ class Mesh
 	Material* material;
 	unsigned int num_indices;
 public:
-	Mesh(std::vector<float>& vertices, std::vector<unsigned int>& indices, std::vector<float>& normals, std::vector<float>& texcoords, char const * pathTexture, char const * specTexture);
+	Mesh(std::vector<float>& vertices, std::vector<uint16_t>& indices, std::vector<float>& normals, std::vector<float>& texcoords, char const * pathTexture, char const * specTexture);
 	~Mesh();
 
 	void Draw();
